// src/constants/enum.ts
export enum RoomLocation {
	NATIONWIDE = 'NATIONWIDE',
	SEOUL = 'SEOUL', // 서울
	GYEONGGI = 'GYEONGGI', // 경기
	INCHEON = 'INCHEON', // 인천
	CHUNGCHEONG = 'CHUNGCHEONG', // 충청
	GYEONGSANG = 'GYEONGSANG', // 경상
	JEOLLA = 'JEOLLA', // 전라
	GANGWON = 'GANGWON', // 강원
	JEJU = 'JEJU', // 제주
	GANGNAM = 'GANGNAM', // 강남
	HONGDAE = 'HONGDAE', // 홍대
	SINCHON = 'SINCHON', // 신촌
	KONKUK = 'KONKUK', // 건대
	DAEHANGNO = 'DAEHANGNO', // 대학로
	OTHER = 'OTHER' // 그 외
}

export enum RoomTheme {
	HORROR = 'HORROR',
	SENTIMENTAL = 'SENTIMENTAL',
	MYSTERY = 'MYSTERY',
	ACTION = 'ACTION',
	SF = 'SF',
	OTHER = 'OTHER'
}